Remove tpmtool, move log functionality to pkg/txtlog
