+++ Gopkg.toml
@@ -24,10 +24,6 @@ ignored = ["github.com/u-root/u-root/bb*"]
   name = "github.com/google/go-tpm"
   revision = "aab8d06b4c87613d2d8f0fcef5d2c9ebed0b6e2e"
 
-[[constraint]]
-  branch = "master"
-  name = "github.com/9elements/tpmtool"
-
 [prune]
   non-go = true
   go-tests = true
