+++ pkg/txtlog/constants.go
@@ -1,38 +1,7 @@
-package tpm
-
-// TPMMaxPCRListSize is the maximum number of PCRs for a TPM
-const TPMMaxPCRListSize = 24
-
-// IAlgHash is the TPM hash algorithm
-type IAlgHash uint16
-
-// We only define TPM hash algorithms here we use
-const (
-	// TPMAlgError is an algorithm error
-	TPMAlgError IAlgHash = 0x0000
-	// TPMAlgSha
-	TPMAlgSha     IAlgHash = 0x0004
-	TPMAlgSha256  IAlgHash = 0x000B
-	TPMAlgSha384  IAlgHash = 0x000C
-	TPMAlgSha512  IAlgHash = 0x000D
-	TPMAlgSm3s256 IAlgHash = 0x0012
-)
-
-// IAlgHashSize is the TPM hash algorithm length
-type IAlgHashSize uint8
-
-const (
-	// TPMAlgShaSize SHA hash size
-	TPMAlgShaSize IAlgHashSize = 20
-	// TPMAlgSha256Size SHA256 hash size
-	TPMAlgSha256Size IAlgHashSize = 32
-	// TPMAlgSha384Size SHA384 hash size
-	TPMAlgSha384Size IAlgHashSize = 48
-	// TPMAlgSha512Size SHA512 hash size
-	TPMAlgSha512Size IAlgHashSize = 64
-	// TPMAlgSm3s256Size SM3-256 hash size
-	TPMAlgSm3s256Size IAlgHashSize = 32
-)
+// Copyright 2020 the u-root Authors. All rights reserved
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+package txtlog
 
 // BIOSLogID is the legacy eventlog type
 type BIOSLogID uint32
@@ -217,28 +186,28 @@ const (
 
 // Txt12LogTypes are the Intel TXT eventlog types
 var TxtLogTypes = map[TxtLogID]string{
-	TxtEvTypeBase: "EVTYPE_BASE",
-	TxtEvTypePcrMapping: "EVTYPE_PCR_MAPPING",
-	TxtEvTypeHashStart: "EVTYPE_HASH_START",
-	TxtEvTypeCombinedHash: "EVTYPE_COMBINED_HASH",
-	TxtEvTypeMleHash: "EVTYPE_MLE_HASH",
-	TxtEvTypeBiosAcRegData: "EVTYPE_BIOSAC_REG_DATA",
-	TxtEvTypeCpuScrtmStat: "EVTYPE_CPU_SCRTM_STAT",
-	TxtEvTypeLcpControlHash: "EVTYPE_LCP_CONTROL_HASH",
-	TxtEvTypeElementsHash: "EVTYPE_ELEMENTS_HASH",
-	TxtEvTypeStmHash: "EVTYPE_STM_HASH",
+	TxtEvTypeBase:               "EVTYPE_BASE",
+	TxtEvTypePcrMapping:         "EVTYPE_PCR_MAPPING",
+	TxtEvTypeHashStart:          "EVTYPE_HASH_START",
+	TxtEvTypeCombinedHash:       "EVTYPE_COMBINED_HASH",
+	TxtEvTypeMleHash:            "EVTYPE_MLE_HASH",
+	TxtEvTypeBiosAcRegData:      "EVTYPE_BIOSAC_REG_DATA",
+	TxtEvTypeCpuScrtmStat:       "EVTYPE_CPU_SCRTM_STAT",
+	TxtEvTypeLcpControlHash:     "EVTYPE_LCP_CONTROL_HASH",
+	TxtEvTypeElementsHash:       "EVTYPE_ELEMENTS_HASH",
+	TxtEvTypeStmHash:            "EVTYPE_STM_HASH",
 	TxtEvTypeOsSinitDataCapHash: "EVTYPE_OSSINITDATA_CAP_HASH",
-	TxtEvTypeSinitPubKeyHash: "EVTYPE_SINIT_PUBKEY_ HASH",
-	TxtEvTypeLcpHash: "EVTYPE_LCP_HASH",
-	TxtEvTypeLcpDetailsHash: "EVTYPE_LCP_DETAILS_HASH",
+	TxtEvTypeSinitPubKeyHash:    "EVTYPE_SINIT_PUBKEY_ HASH",
+	TxtEvTypeLcpHash:            "EVTYPE_LCP_HASH",
+	TxtEvTypeLcpDetailsHash:     "EVTYPE_LCP_DETAILS_HASH",
 	TxtEvTypeLcpAuthoritiesHash: "EVTYPE_LCP_AUTHORITIES_HASH",
-	TxtEvTypeNvInfoHash: "EVTYPE_NV_INFO_HASH",
-	TxtEvTypeColdBootBiosHash: "EVTYPE_COLD_BOOT_BIOS_HASH",
-	TxtEvTypeKmHash: "EVTYPE_KM_HASH",
-	TxtEvTypeBpmHash: "EVTYPE_KM_HASH",
-	TxtEvTypeKmInfoHash: "EVTYPE_KM_INFO_HASH",
-	TxtEvTypeBpmInfoHash: "EVTYPE_BPM_INFO_HASH",
-	TxtEvTypeBootPolHash: "EVTYPE_BOOT_POL_HASH",
-	TxtEvTypeRandValue: "EVTYPE_RANDOM_VALUE",
-	TxtEvTypeCapValue: "EVTYPE_CAP_VALUE",
+	TxtEvTypeNvInfoHash:         "EVTYPE_NV_INFO_HASH",
+	TxtEvTypeColdBootBiosHash:   "EVTYPE_COLD_BOOT_BIOS_HASH",
+	TxtEvTypeKmHash:             "EVTYPE_KM_HASH",
+	TxtEvTypeBpmHash:            "EVTYPE_KM_HASH",
+	TxtEvTypeKmInfoHash:         "EVTYPE_KM_INFO_HASH",
+	TxtEvTypeBpmInfoHash:        "EVTYPE_BPM_INFO_HASH",
+	TxtEvTypeBootPolHash:        "EVTYPE_BOOT_POL_HASH",
+	TxtEvTypeRandValue:          "EVTYPE_RANDOM_VALUE",
+	TxtEvTypeCapValue:           "EVTYPE_CAP_VALUE",
 }
